--- src/main/java/com/mobble/mobbleserver/refactor/chat/chatRoom/entity/ChatRoomType.java
@@ -1,6 +1,0 @@
-package com.mobble.mobbleserver.refactor.chat.chatRoom.entity;
-
-public enum ChatRoomType {
-    GROUP,
-    DIRECT;
-}
