+++ src/main/java/com/mobble/mobbleserver/refactor/chat/chatRoom/repository/ChatRoomRepository.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.refactor.chat.chatRoom.repository;
 
-import com.mobble.mobbleserver.refactor.chat.chatRoom.entity.ChatRoom;
+import com.mobble.mobbleserver.domain.chat.room.ChatRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
