+++ src/main/java/com/mobble/mobbleserver/refactor/chat/chatRoom/validator/ChatRoomValidator.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.refactor.chat.chatRoom.validator;
 
-import com.mobble.mobbleserver.refactor.chat.chatRoom.entity.ChatRoom;
+import com.mobble.mobbleserver.domain.chat.room.ChatRoom;
 import com.mobble.mobbleserver.refactor.chat.chatRoom.repository.ChatRoomRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
