+++ src/main/java/com/mobble/mobbleserver/refactor/chat/chatRoomParticipant/repository/ChatRoomParticipantRepository.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.refactor.chat.chatRoomParticipant.repository;
 
-import com.mobble.mobbleserver.refactor.chat.chatRoomParticipant.entity.ChatRoomParticipant;
+import com.mobble.mobbleserver.domain.chat.room.ChatRoomParticipant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
