+++ src/main/java/com/mobble/mobbleserver/refactor/clubMember/repository/ClubMemberRepositoryImpl.java
@@ -6,7 +6,7 @@ import lombok.RequiredArgsConstructor;
 
 import java.util.List;
 
-import static com.mobble.mobbleserver.refactor.chat.chatRoom.entity.QChatRoom.chatRoom;
+import static com.mobble.mobbleserver.domain.chat.room.QChatRoom.chatRoom;
 import static com.mobble.mobbleserver.refactor.chat.clubChatRoom.entity.QClubChatRoom.clubChatRoom;
 import static com.mobble.mobbleserver.refactor.club.core.entity.QClub.club;
 import static com.mobble.mobbleserver.refactor.clubMember.entity.QClubMember.clubMember;
