+++ src/test/java/com/mobble/mobbleserver/support/fixture/chat/chatRoom/ChatRoomTestFixture.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.support.fixture.chat.chatRoom;
 
-import com.mobble.mobbleserver.refactor.chat.chatRoom.entity.ChatRoom;
-import com.mobble.mobbleserver.refactor.chat.chatRoom.entity.ChatRoomType;
+import com.mobble.mobbleserver.domain.chat.room.ChatRoom;
+import com.mobble.mobbleserver.domain.chat.room.ChatRoomType;
 
 public final class ChatRoomTestFixture {
     private ChatRoomTestFixture() {}
