+++ src/test/java/com/mobble/mobbleserver/support/fixture/chat/clubChatRoom/ClubChatRoomTestFixture.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.support.fixture.chat.clubChatRoom;
 
-import com.mobble.mobbleserver.refactor.chat.chatRoom.entity.ChatRoom;
+import com.mobble.mobbleserver.domain.chat.room.ChatRoom;
 import com.mobble.mobbleserver.refactor.chat.clubChatRoom.entity.ClubChatRoom;
 import com.mobble.mobbleserver.refactor.club.core.entity.Club;
 import com.mobble.mobbleserver.support.fixture.chat.chatRoom.ChatRoomTestFixture;
